Fix float double type error

+++ src/framework/models/account/Entry.java
@@ -6,10 +6,10 @@ public class Entry implements IEntry{
 	private double amount;
 	private LocalDate date;
 
-	public Float getAmount() {
+	public double getAmount() {
 		return amount;
 	}
-	public void setAmount(Float amount) {
+	public void setAmount(double amount) {
 		this.amount = amount;
 	}
 	public LocalDate getDate() {
